Update value of user button in real time
The button in the top bar that was displaying the name of the user was not
being updated timely when the user logged in. To solve this issue,
I introduced a computed property retrieving the username value from the
store and added the created hook that updates the value of the button
as soon as the computed property is evaluated.