Add error message and routing in login form
The login form now shows an error message when invalid credentials are
inserted. I implemented this adding a new variable in the login store
that contains the error message and is displayed when it is not empty.

The routing from the login page to the homepage was happening ahead of
authentication. I solved the issue adding an instance of the router in the
login store and routing to the homepage in case of successful
authentication.