Create user data modification form
In the personal data section, it is now possible to switch to a form
where fields can be edited. This implies the following changes:
- adding actions to mutate the state of the application (a proper call
to Strapi API is still needed to actually change data in the backend);
- creating a new component (ModifyUserDataCard) that enables the input
of new values;
- enabling a switch logic in the ProfileIndex that makes the components
dynamically alternate in the view depending on user interaction;
- modifying the UserDataCard adding a button to enable the modification
of the fields.